server-main/requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
Markdown==3.3.7
numpy==1.23.0
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
